power: qpnp-smbcharger-mmi: Add Suspend SOC Recalc
When the System Suspends the running last reported SOC needs to
be reset.  The reason is that the SOC could have moved a large
amount while sleeping.  The curretn SOC needs to be recalculated
so that the changes is noticable to the User upon wakeup.

To do this keep track of the Suspend state as well as the fact
that the heartbeat has not run yet after a Suspend.  The
WAS_SUSPEND bit will track this and then be cleared once the
heartbeat gets past the point that the information is not longer
needed.  For Dual Batteries, this will force the last_reported_soc
to be updated with the current calculation rather then trying
to Monotonically adjust it slowly.

Change-Id: I0de830df74720f0f26f8f7a7cfcba7698f6a6dd6
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: https://gerrit.mot.com/1355819
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Curtis Porter <porterc@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Submit-Approved: Jira Key