Change dt-gesture sensor to SPECIAL_REPORTING mode
dt-gesture sensor was previously an ON_CHANGE sensor. This is
generally okay, but it can lead to the Sensor Service sending out
an old event to a new (second, third, etc..) client if there was
already a previous client registered for the sensor, and this is
not the behavior that is desired.

Change-Id: I3bc91c5271fedc83b11ee5028980cfea99dd6732
Reviewed-on: https://gerrit.mot.com/1484754
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Konstantin Makariev <kmakariev@motorola.com>
Submit-Approved: Jira Key