aw8695: Cancel the 120s timer when stopping vibrating
When stopping waveform playing, the 120s timer to safely stop
vibrating also needs to be canceled because it becomes no use, and
it might disturb normal operation when it's triggered later.

NO_PROP_NEEDED: This propagation is only needed when the change of
framework for this new feature is integrated, being done by engineers
in charge of that product.

Change-Id: I5f5b65dfc930d252c89183c2bf22f7b8cae110aa
Signed-off-by: yangjq <yangjq@lenovo.com>
Reviewed-on: https://gerrit.mot.com/1613915
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Wentao Xu <xuwt2@motorola.com>
Reviewed-by: Jianqi Yang <yangj@motorola.com>
Submit-Approved: Jira Key