aw8697: A fix of rtp playing from Awinic
When frequently making a phone call and declining it, the waveform
playing sometimes would fail to work due to the limitation of i2c's
transmission speed.

Awinic provides this solution and it not only works fine but also has
been integrated to all their new software releases, even for aw8697.

NO_PROP_NEEDED: This propagation is only needed when the change of
framework for this new feature is integrated, being done by engineers
in charge of that product.

Change-Id: Id7f1a01c4fcb13b082ca48d105654803eb84e8f4
Signed-off-by: yangjq <yangjq@lenovo.com>
Reviewed-on: https://gerrit.mot.com/1616585
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Jianqi Yang <yangj@motorola.com>
Submit-Approved: Jira Key